Day 10
## PR Type 🥰
- 🥰 기능 추가 또는 삭제
- [ ] 버그 수정
- [ ] 주석추가 또는 수정
- [] 문서 수정
- [ ] 코드 리펙토링
- [ ] 코드에 영향을 주지 않는 변경사항
- [ ] 의존성, 환경 변수, 빌드 관련 코드 업데이트

## 주요 변경 사항 🔑
- WaferMap 작성

+++ Klarf_pyj2/Model/WaferModel.cs
@@ -10,7 +10,7 @@ namespace Klarf
 {
     public class WaferModel
     {
-        public double width, height;
+        public int width, height;
         public List<string> xIndex, yIndex;
         public List<int> xIndices, yIndices;
